move courseListView from fragments to custom
